Rename Mutable to Writable
Closes #396
