+++ test-d/writable.ts
@@ -1,0 +1,28 @@
+import {expectType, expectError} from 'tsd';
+import type {Writable} from '../index';
+
+type Foo = {
+	readonly a: number;
+	readonly b: string;
+};
+
+const ab: Writable<Foo> = {a: 1, b: '2'};
+ab.a = 2;
+const ab2: Writable<Readonly<Foo>> = ab;
+ab2.a = 2;
+
+// Update one writable and one readonly to writable, leaving one property unaffected.
+declare const variation1: Writable<{readonly a: number; b: string; readonly c: boolean}, 'b' | 'c'>;
+expectType<{readonly a: number; b: string; c: boolean}>(variation1);
+
+// Update two readonly to writable, leaving one property unaffected.
+declare const variation2: Writable<{readonly a: number; readonly b: string; readonly c: boolean}, 'a' | 'b'>;
+expectType<{a: number; b: string; readonly c: boolean}>(variation2);
+
+// Three writable remain writable.
+declare const variation3: Writable<{a: number; b: string; c: boolean}, 'a' | 'b' | 'c'>;
+expectType<{a: number; b: string; c: boolean}>(variation3);
+
+// Check if type changes raise an error even if readonly and writable are applied correctly.
+declare const variation4: Writable<{readonly a: number; b: string; readonly c: boolean}, 'b' | 'c'>;
+expectError<{readonly a: boolean; b: string; c: boolean}>(variation4);
